cleanup per Matt
