+++ src/models/hooks/goal.test.js
@@ -242,36 +242,6 @@ describe('goal hooks', () => {
     });
   });
 
-  describe('goalSourceValidations', () => {
-    const sequelize = {};
-    const instance = {
-      changed: jest.fn(),
-      set: jest.fn(),
-    };
-    const options = {
-      fields: [],
-    };
-
-    afterEach(() => jest.clearAllMocks());
-    it('throws an error for invalid tr source', () => {
-      instance.changed.mockReturnValueOnce(['source']);
-      instance.source = 'invalid_tr_source';
-      instance.createdVia = 'tr';
-      expect(() => {
-        onlyAllowTrGoalSourceForGoalsCreatedViaTr(sequelize, instance, options);
-      }).toThrowError('Goals created via a Training Report must have a source of "Training event".');
-    });
-
-    it('should not throw an error for valid tr source', () => {
-      instance.changed.mockReturnValueOnce(['source']);
-      instance.source = 'Training event';
-      instance.createdVia = 'tr';
-      expect(() => {
-        onlyAllowTrGoalSourceForGoalsCreatedViaTr(sequelize, instance, options);
-      }).not.toThrowError();
-    });
-  });
-
   describe('findOrCreateGoalTemplate', () => {
     it('should find or create a goal template', async () => {
       const sequelize = {
